requirements/base.txt
aiohttp==3.12.15
asgiref==3.9.1
click==8.2.1
fastapi==0.116.1
filetype==1.2.0
jinja2==3.1.6
nest-asyncio==1.6.0
pydantic==2.11.3
python-multipart==0.0.20
python-telegram-bot==22.3
starlette==0.47.3
uvicorn==0.35.0